Update ler_dados_int return type to usize and refactor client data input handling

+++ src/tela/ler.rs
@@ -6,7 +6,7 @@ pub fn ler_dados() -> String {
     dados.trim().to_string()
 }
 
-pub fn ler_dados_int() -> i32 {
+pub fn ler_dados_int() -> usize {
     let mut dados: String = String::new();
     io::stdin().read_line(&mut dados).expect("Falha ao ler a entrada");
     dados.trim().parse().expect("Erro ao converer dados para inteiro")
